# api/openapi.yaml
openapi: 3.0.3
info:
  title: My Hike API
  version: 1.0.0
  description: |
    REST API for My Hike. Except for registration, login, health, and this
    document, endpoints require a JWT access token in the Authorization header.
    Administrator endpoints additionally require the `admin` role.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/weights:
    get:
      tags: [Weights]
      summary: List the authenticated user's weight measurements
      operationId: listWeights
      responses:
        "200":
          description: Weight measurements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Weight"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Weights]
      summary: Record body weight
      operationId: createWeight
      requestBody:
        $ref: "#/components/requestBodies/WeightInput"
      responses:
        "201":
          description: Weight recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Weight"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/weights/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Weights]
      summary: Get an owned weight measurement; administrators may read any measurement
      operationId: getWeight
      responses:
        "200":
          description: Weight measurement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Weight"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Weights]
      summary: Update an owned weight measurement; administrators may update any measurement
      operationId: updateWeight
      requestBody:
        $ref: "#/components/requestBodies/WeightInput"
      responses:
        "200":
          description: Updated weight measurement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Weight"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Weights]
      summary: Delete an owned weight measurement; administrators may delete any measurement
      operationId: deleteWeight
      responses:
        "204":
          description: Weight measurement deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/weights/migrate:
    post:
      tags: [Weights]
      summary: Migrate browser weight records into the authenticated account
      operationId: migrateWeights
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/WeightMigrationItem"
      responses:
        "200":
          description: Migration counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrationResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  requestBodies:
...
    WeightInput:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WeightInput"
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, expired, or deleted-user token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found or not owned by the authenticated user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    WeightInput:
      type: object
      required: [weightKg, recordedOn]
      properties:
        weightKg:
          type: number
          format: double
          minimum: 20
          maximum: 500
        recordedOn:
          type: string
          format: date
    WeightMigrationItem:
      allOf:
        - $ref: "#/components/schemas/WeightInput"
        - type: object
          required: [id]
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
    Weight:
      allOf:
        - $ref: "#/components/schemas/WeightMigrationItem"
        - type: object
          required: [userId]
          properties:
            userId:
              type: string
              format: uuid
...
    MigrationResult:
      type: object
      required: [imported, skipped]
      properties:
        imported:
          type: integer
          minimum: 0
        skipped:
          type: integer
          minimum: 0
...
security:
  - bearerAuth: []